Simplify JfrRuntimeAccess
