+++ substratevm/src/com.oracle.svm.jfr/src/com/oracle/svm/jfr/JfrFeature.java
@@ -94,7 +94,7 @@ public class JfrFeature implements Feature {
         ImageSingletons.add(SubstrateJVM.class, new SubstrateJVM());
         ImageSingletons.add(JfrManager.class, new JfrManager());
         ImageSingletons.add(JfrSerializerSupport.class, new JfrSerializerSupport());
-        ImageSingletons.add(JfrRuntimeAccess.class, new JfrRuntimeAccessImpl());
+        ImageSingletons.add(JfrRuntimeAccess.class, new JfrRuntimeAccess());
 
         JfrSerializerSupport.get().register(new JfrFrameTypeSerializer());
         ThreadListenerSupport.get().register(SubstrateJVM.getThreadLocal());
